% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_logic.R
\name{acd4_stat_}
\alias{acd4_stat_}
\title{acd4_stat_}
\usage{
acd4_stat_(d_nterm)
}
\arguments{
\item{d_nterm}{input data frame}
}
\value{
list of dataframes.For each group , cTIS and aTIS dataframe are included.
}
\description{
Process statistics for ACd4 low confident
}
\author{
Andrea Argentini
}
