% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_logic.R
\name{acetyl_stat_}
\alias{acetyl_stat_}
\title{acetyl_stat_}
\usage{
acetyl_stat_(d_nterm)
}
\arguments{
\item{d_nterm}{input data frame}
}
\value{
list of dataframes.For each group , cTIS and aTIS dataframe are included.
}
\description{
Process statistics for percentage of acetylation
}
\author{
Andrea Argentini
}
