% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_logic.R
\name{annotate_df}
\alias{annotate_df}
\title{annotate_df}
\usage{
annotate_df(res, type, groups)
}
\arguments{
\item{res}{A nested list of data frames containing proteomics results.}

\item{type}{Character string; the sub-list key to access (e.g., "p_start_").}

\item{groups}{A character vector of group labels to compare.}
}
\value{
The modified `res` list with the added `Venn` column in the specified sub-data frames.
}
\description{
This function annotates a list of data frames with a "Venn" column indicating 
in which experimental groups each peptide sequence was identified.
}
\author{
Andrea Argentini
}
