% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_logic.R
\name{check_columns_presence}
\alias{check_columns_presence}
\title{check_columns_presence}
\usage{
check_columns_presence(df, min_features)
}
\arguments{
\item{df}{Input data frame to be checked.}

\item{min_features}{Character vector of required column names.}
}
\value{
A named list containing:
\itemize{
  \item \code{status}: Integer; 0 for success, 1 if columns are missing.
  \item \code{error}: Character string containing the error message and missing columns.
}
}
\description{
This helper function checks for the presence of a required set 
of column names in the input data frame.
}
\author{
Andrea Argentini
}
