% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_logic.R
\name{check_dependencies}
\alias{check_dependencies}
\title{check_dependencies}
\usage{
check_dependencies(required_packages = required_packages)
}
\arguments{
\item{required_packages}{list with packages to be installed}
}
\value{
none
}
\description{
This function gets a vector with names of packages and it
Check if packages are available, if not it installs them.
Otherwise, it loads the  requiredpackages.
}
\author{
Andrea Argentini
}
