% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_logic.R
\name{create_set_indicators}
\alias{create_set_indicators}
\title{create_set_indicators}
\usage{
create_set_indicators(sets)
}
\arguments{
\item{sets}{}
}
\value{
none
}
\description{
Function to create a visual representation of set membership
}
\author{
andrea argentini
}
