% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_logic.R
\name{get_all_intersections}
\alias{get_all_intersections}
\title{get_all_intersections}
\usage{
get_all_intersections(df)
}
\arguments{
\item{df}{A data.frame with one "Elements" column and one or more logical columns.}
}
\value{
A named list where each entry corresponds to a non-empty intersection.
}
\description{
Computes all non-empty intersections of logical set membership columns 
in a given data frame.
}
\author{
andrea argentini
}
