% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_logic.R
\name{global_PEP_general}
\alias{global_PEP_general}
\title{global_PEP_general}
\usage{
global_PEP_general(pep_nterm, design)
}
\arguments{
\item{d_nterm}{input n-terminal data frame}
}
\value{
res dataframe with all the computed metrics
}
\description{
Place holder function for peptide quant processing. 
Numb identification 
Acetilated %
}
\author{
Andrea Argentini
}
