% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_logic.R
\name{global_PSM_general}
\alias{global_PSM_general}
\title{global_PSM_general}
\usage{
global_PSM_general(d_nterm, stat_reg, stat_name)
}
\arguments{
\item{d_nterm}{A data frame containing PSM-level N-terminal data.}

\item{stat_reg}{A list of regular expressions defining the modifications to search for.}

\item{stat_name}{A character vector of names corresponding to the `stat_reg` patterns.}
}
\value{
A list containing:
\itemize{
  \item{error}{ Character string containing error messages, if any.}
  \item{status}{ Integer status code (0 for success, 1 for error).}
  \item{res}{ Dataframe of global summary statistics.}
  \item{res_sample}{ Dataframe of summary statistics broken down by sample.}
}
}
\description{
Computes global PTM statistics from PSM data using a list of regular expressions. 
Calculates specific proteomics metrics including SCX enrichment efficiency 
and N-terminal acetylation percentages across the entire dataset and per sample.
}
\author{
Andrea Argentini
}
