% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_logic.R
\name{group_stat}
\alias{group_stat}
\title{group_stat}
\usage{
group_stat(d_nterm)
}
\arguments{
\item{d_nterm}{input n-terminal data frame}
}
\value{
res dataframe with all the computed statistics

l_uniq_protein list of all the unique protein per group
}
\description{
Compute statistics at group level. Statistic and its labels are
parameters at the moment.
}
\author{
Andrea Argentini
}
