% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_logic.R
\name{process_dataset}
\alias{process_dataset}
\title{process_dataset}
\usage{
process_dataset(dataset)
}
\arguments{
\item{dataset}{input  data frame}
}
\value{
Data frame with aggregatted information
}
\description{
Add futher aggregated metrics from acytilated and ACd4 statistics.
}
\author{
Lode
}
