% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_logic.R
\name{process_nterm_data}
\alias{process_nterm_data}
\title{process_nterm_data}
\usage{
process_nterm_data(params_report, analysis_type = "id_")
}
\arguments{
\item{params_report}{A named list containing input parameters (input_file, design_file, select_group).}

\item{analysis_type}{Character string specifying the logic branch. Defaults to "id_".}
}
\value{
A named list containing `quarto_data` and `export_data`.
}
\description{
Internal helper function that orchestrates the data analysis pipeline. It reads the raw input,
applies the statistical logic based on the `analysis_type`, and structures the results into
two distinct "data bags".
}
\author{
andrea argentini
}
\keyword{internal}
