% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_logic.R
\name{quant_base_}
\alias{quant_base_}
\title{quant_base_}
\usage{
quant_base_(d_nterm)
}
\arguments{
\item{d_nterm}{input data frame}
}
\value{
list of dataframes.For each group , cTIS and aTIS dataframe are included.
}
\description{
quant base dataframe per group
}
\author{
Andrea Argentini
}
