% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_template.R
\name{render_nterm_report}
\alias{render_nterm_report}
\title{Render a DDA report using a Quarto template}
\usage{
render_nterm_report(
  params_report,
  template_ids,
  report_folder,
  report_filename
)
}
\arguments{
\item{params_report}{parameters}

\item{template_ids}{template ids can be a vector}

\item{report_folder}{description}

\item{report_filename}{output report file}
}
\value{
The full path to the rendered report.
}
\description{
Render a DDA report using a Quarto template
}
\details{
The `params` list must contain the following elements:
\describe{
  \item{\code{test}}{test}
  \item{\code{test}}{test}

}
}
\author{
andrea Argentini
}
