% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_template.R
\name{render_nterm_website}
\alias{render_nterm_website}
\title{render_nterm_website}
\usage{
render_nterm_website(
  params_dda,
  params_dia,
  template_dda,
  report_folder,
  template_dia,
  project_id
)
}
\arguments{
\item{params_dda}{A named list of parameters to pass to the N-Term (DDA) report rendering function.}

\item{params_dia}{A named list of parameters to pass to the DIA report rendering function.}

\item{template_dda}{Path to the Quarto template (.qmd) for the N-Term (DDA) report.}

\item{report_folder}{Directory where the final rendered website will be copied. Should be writable.}

\item{template_dia}{Path to the Quarto template (.qmd) for the DIA report.}

\item{project_id}{@return
Invisibly returns the path to the final website in \code{report_folder}.}
}
\description{
render_nterm_website
}
\author{
andrea Argentini
}
