% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_template.R
\name{render_ntermdia_report}
\alias{render_ntermdia_report}
\title{render_dia_report}
\usage{
render_ntermdia_report(
  params_report,
  template_file,
  report_target_folder,
  report_filename
)
}
\arguments{
\item{params_report}{List of parameters for the report.}

\item{template_file}{Path to the .qmd template.}

\item{report_target_folder}{Directory to save the output.}

\item{report_filename}{Name of the output file}
}
\description{
render_dia_report
}
\author{
andrea Argentini
}
