% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_template.R
\name{validate_filename}
\alias{validate_filename}
\title{Validate file name}
\usage{
validate_filename(filename)
}
\arguments{
\item{filename}{The file name to be validated.}
}
\value{
TRUE if the file name is valid, otherwise stops with an error message.
}
\description{
Validate file name
}
