% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_template.R
\name{validate_folder}
\alias{validate_folder}
\title{Validate report folder path}
\usage{
validate_folder(report_folder)
}
\arguments{
\item{report_folder}{The folder path to be validated.}
}
\value{
TRUE if the folder path is valid and writable, otherwise stops with an error message.
}
\description{
Validate report folder path
}
