% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_template.R
\name{validate_template}
\alias{validate_template}
\title{validate_template}
\usage{
validate_template(t_id)
}
\arguments{
\item{t_id}{Character string representing the template ID.}
}
\value{
The filename (string) of the corresponding Quarto template.
}
\description{
Maps a user-friendly template ID to an actual .qmd file name.
}
