% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_logic.R
\name{annotate_df}
\alias{annotate_df}
\title{annotate_df}
\usage{
annotate_df(res, type, groups)
}
\arguments{
\item{res}{list with packages to be installed}

\item{groups}{vector of group label}
}
\value{
none
}
\description{
This function annotate the df with N groups
}
\author{
Andrea Argentini
}
