% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_logic.R
\name{get_all_intersections}
\alias{get_all_intersections}
\title{get_all_intersections}
\usage{
get_all_intersections(df)
}
\arguments{
\item{df}{}
}
\value{
results
}
\description{
Modified get_all_intersections function to include additional metadata
}
\author{
andrea argentini
}
