% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_logic.R
\name{global_PEP_general}
\alias{global_PEP_general}
\title{global_PEP_general}
\usage{
global_PEP_general(pep_nterm, design)
}
\arguments{
\item{pep_nterm}{input n-terminal data frame parsed from peptide file}

\item{design}{dataframe realted to the design experiment}
}
\value{
A list containing:
\itemize{
  \item{error} Error message string.
  \item{status} Status code (0 for success).
  \item{pep_cnt_sample} Precursor counts per sample.
  \item{pep_cnt_group} Precursor counts per group.
  \item{ace_group} Acetylated peptides split by group.
  \item{ace_sample} Acetylated peptides split by sample.
}
}
\description{
Calculates the number of identified precursors per Sample and Group,
and filters for N-terminally acetylated peptides based on specific 
methionine (M) cleavage logic
}
\author{
Andrea Argentini
}
