% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_logic.R
\name{global_PSM_general}
\alias{global_PSM_general}
\title{global_PSM_general}
\usage{
global_PSM_general(d_nterm, stat_reg, stat_name)
}
\arguments{
\item{d_nterm}{input n-terminal data frame}
}
\value{
res dataframe with all the computed metrics
}
\description{
Global statistics computed for the entire dataset.Statistic and its labels are
parameters at the moment.
}
\author{
Andrea Argentini
}
