% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_logic.R
\name{process_filter_wip}
\alias{process_filter_wip}
\title{process_filter_wip}
\usage{
process_filter_wip(filter_item, filter_label, data, task)
}
\arguments{
\item{filter_item}{A list containing the filtering logic (e.g., `filter_item$logic`).}

\item{filter_label}{A character string used as a label for the specific metric.}

\item{data}{The input data frame to be filtered.}

\item{task}{An array or vector of mascot tasks used for context (if applicable within the logic).}
}
\value{
A list containing:
\itemize{
  \item{filter_name}{ The provided `filter_label`.}
  \item{val_count}{ The number of rows (integer) after applying the filter.}
}
}
\description{
Applies a specific filtering logic to the input data frame and computes 
the absolute count of the remaining rows. This function supports 
tidy evaluation for dynamic filtering.
}
\author{
Andrea Argentini
}
