% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_logic.R
\name{process_filter_wip}
\alias{process_filter_wip}
\title{process_filter_wip}
\usage{
process_filter_wip(filter_item, filter_label, data, countpeptides, task)
}
\arguments{
\item{filter_label}{label of the filter / metrics}

\item{data}{inpute dataframe to filter}

\item{countpeptides}{denominator cardinality of the input dataframe}

\item{filter_pattern}{reg exp to filter}
}
\value{
list with computed metrics
}
\description{
This function apply the right filtering and computed the metrics related,
bot as absolute and percentage values.
}
\author{
Andrea Argentini
}
