% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_logic.R
\name{read_data}
\alias{read_data}
\title{read_data}
\usage{
read_data(file_nterm, file_expdesign)
}
\arguments{
\item{file_nterm}{Path to input mascot report}

\item{file_expdesign}{Path to design TSV file}
}
\value{
Data frame with imported data
}
\description{
It imports the mascot PSM report and the experiments design file.
}
\author{
Andrea Argentini
}
