% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_logic.R
\name{render_child}
\alias{render_child}
\title{render_child}
\usage{
render_child(data, path, label, template)
}
\arguments{
\item{data}{DE result for each comparison}

\item{path}{PAth where to store the result}

\item{label}{layer name of layer in the qfeature object}

\item{template}{name of the template .Rms file (contrast,heatmap etc)}
}
\value{
none
}
\description{
This function allows to render other template.Rmd in the main quarto document
}
\author{
andrea argentini
}
