% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_template.R
\name{render_ntermdia_report}
\alias{render_ntermdia_report}
\title{render_dia_report}
\usage{
render_ntermdia_report(
  params_report,
  template_file,
  report_target_folder,
  report_filename
)
}
\description{
render_dia_report
}
\author{
andrea Argentini
}
