% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_template.R
\name{render_quarto_index}
\alias{render_quarto_index}
\title{render_quarto_index}
\usage{
render_quarto_index(report_info, out_file = "", title = "")
}
\arguments{
\item{report_info}{report}

\item{out_file}{output file}

\item{title}{title of the}
}
\description{
render_quarto_index
}
\author{
andrea Argentini
}
