% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_template.R
\name{validate_template}
\alias{validate_template}
\title{Validate template parameter}
\usage{
validate_template(template)
}
\arguments{
\item{template}{The template parameter to be validated.}
}
\value{
TRUE if the template is valid, otherwise stops with an error message.
}
\description{
Validate template parameter
}
