% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_logic.R
\name{view_intersection}
\alias{view_intersection}
\title{view_intersection}
\usage{
view_intersection(df, sets = list(), exclusive = TRUE)
}
\arguments{
\item{df}{}

\item{sets}{}

\item{exclusive}{}
}
\value{
none
}
\description{
Helper function to view intersection (same as before)
}
\author{
andrea argentini
}
