% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_logic.R
\name{write_final_result}
\alias{write_final_result}
\title{write_final_result}
\usage{
write_final_result(acetyl, acd4, quant_base, group_l, path)
}
\arguments{
\item{required_packages}{list with packages to be installed}
}
\value{
int
}
\description{
This function save all result in a excell file
}
\author{
Andrea Argentini
}
